web_app/backend/services/srne_adaptor/requirements.txt
redis
pymodbus==3.9.2
pyserial
psycopg2
paho-mqtt